Remove the concept of legacy/latest detail pages
