+++ shell/models/secret.js
@@ -625,4 +625,8 @@ export default class Secret extends SteveModel {
 
     return this.t('resourceTable.groupLabel.project', { name: escapeHtml(this?.project?.nameDisplay || '') }, true);
   }
+
+  get fullDetailPageOverride() {
+    return true;
+  }
 }
