--- shell/pages/explorer/resource/detail/projectsecret.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-
-import Secret from '@shell/pages/explorer/resource/detail/secret.vue';
-
-</script>
-
-<template>
-  <Secret :isProjectSecret="true" />
-</template>
